Modify the changeableTo method's method signature in all implementations. Change the return value from Map<Animal, Integer> to Map<Animal, Double>.

+++ src/main/java/hu/eszter/bokkon/model/animal/Animal.java
@@ -4,6 +4,6 @@ import java.util.Map;
 
 public interface Animal {
 
-   Map<Animal, Integer> changeableTo();
+    Map<Animal, Double> changeableTo();
 
 }
