+++ src/main/java/hu/eszter/bokkon/model/animal/Wolf.java
@@ -11,7 +11,7 @@ public final class Wolf implements AttackerAnimal {
     }
 
     @Override
-    public Map<Animal, Integer> changeableTo() {
+    public Map<Animal, Double> changeableTo() {
         return new HashMap<>();
     }
 }
